+++ packages/remix-node/sessions/fileStorage.ts
@@ -27,7 +27,7 @@ interface FileSessionStorageOptions {
  * The advantage of using this instead of cookie session storage is that
  * files may contain much more data than cookies.
  *
- * @see https://remix.run/api/remix#createfilesessionstorage-node
+ * @see https://remix.run/utils/sessions#createfilesessionstorage-node
  */
 export function createFileSessionStorage({
   cookie,