+++ packages/remix-server-runtime/sessions/memoryStorage.ts
@@ -24,7 +24,7 @@ export type CreateMemorySessionStorageFunction = (
  * Note: This storage does not scale beyond a single process, so it is not
  * suitable for most production scenarios.
  *
- * @see https://remix.run/api/remix#creatememorysessionstorage
+ * @see https://remix.run/utils/sessions#creatememorysessionstorage
  */
 export const createMemorySessionStorageFactory =
   (